refactor(mention): extract suggestion item schema
