+++ src/lib/actions/autocomplete.svelte.ts
@@ -2,8 +2,8 @@ import { nip19 } from 'nostr-tools';
 import { mount, unmount } from 'svelte';
 import { browser } from '$app/environment';
 import MentionMenu from '$lib/components/MentionMenu.svelte';
+import type { MentionItem } from '$lib/mention';
 import { profileSuggestions } from '$lib/stores/profileSuggestions';
-import type { MentionItem } from '$lib/types';
 
 // This intentionally hand-rolls keyboard navigation instead of using
 // @skeletonlabs/skeleton-svelte's `Menu` or `Combobox` (both wrap @zag-js
