+++ src/lib/actions/autocomplete.test.ts
@@ -2,8 +2,8 @@ import { fireEvent, render } from '@testing-library/svelte';
 import { nip19 } from 'nostr-tools';
 import { writable } from 'svelte/store';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import type { MentionItem } from '$lib/mention';
 import type { ProfileSuggestions, ProfileSuggestionsState } from '$lib/stores/profileSuggestions';
-import type { MentionItem } from '$lib/types';
 
 const { profileSuggestionsMock } = vi.hoisted(() => ({ profileSuggestionsMock: vi.fn() }));
